Share the git smart-HTTP user principal glue between route groups
Asset and agent-state smart-HTTP routes carried identical copies of the
verdict resolution, principal construction, and substrate adapter
helpers. The shared module parameterizes the verdict resource string,
so the security-critical authz boundary lives in one place. Behavior,
denial strings, and route logic are unchanged.
